# docs/api/openapi.yaml
openapi: 3.0.3

info:
  title: Asturias en Familia — API REST
  description: |
    API REST del proyecto Intermodular II · DAW · IES Nº1 Gijón · 2024-2025.
    Proporciona acceso a actividades familiares geolocalizadas en Asturias.
  version: 1.0.0

servers:
  - url: http://localhost:3000
    description: Desarrollo local (Docker)

tags:
  - name: actividades
    description: Gestión de actividades familiares
  - name: categorías
    description: Categorías de actividades
  - name: auth
    description: Autenticación y gestión de sesión

# ─── Componentes reutilizables ──────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    # ── Paginación ──────────────────────────────────────────
    Pagination:
      type: object
      properties:
        total:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        pages:
          type: integer
          example: 3

    # ── Categoría ───────────────────────────────────────────
    Category:
      type: object
      properties:
        _id:
          type: string
          example: "664a1b2c3d4e5f6a7b8c9d0e"
        name:
          type: string
          example: "Rutas"
        slug:
          type: string
          example: "rutas"
        icon:
          type: string
          example: "icon-rutas.svg"
        color:
          type: string
          example: "#3D7A36"
        description:
          type: string
          example: "Rutas de senderismo y naturaleza para toda la familia"
        order:
          type: integer
          example: 1
        active:
          type: boolean
          example: true

    CategoryInput:
      type: object
      required: [name, slug, icon, color]
      properties:
        name:
          type: string
          maxLength: 50
          example: "Rutas"
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
          example: "rutas"
        icon:
          type: string
          example: "icon-rutas.svg"
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          example: "#3D7A36"
        description:
          type: string
          maxLength: 200
        order:
          type: integer
          minimum: 0

    # ── Actividad ───────────────────────────────────────────
    Activity:
      type: object
      properties:
        _id:
          type: string
          example: "664a1b2c3d4e5f6a7b8c9d0f"
        name:
          type: string
          example: "Ruta del Oso — Picos de Europa"
        description:
          type: string
          example: "Ruta circular de dificultad baja ideal para familias con niños mayores de 6 años."
        category:
          $ref: '#/components/schemas/Category'
        location:
          type: object
          properties:
            type:
              type: string
              enum: [Point]
            coordinates:
              type: array
              items:
                type: number
              minItems: 2
              maxItems: 2
              description: "[longitud, latitud] — orden GeoJSON"
              example: [-4.9902, 43.2547]
        zone:
          type: string
          enum: [oriente, centro, occidente]
          example: "oriente"
        municipality:
          type: string
          example: "Cangas de Onís"
        images:
          type: array
          items:
            type: string
          example: ["https://example.com/img1.jpg"]
        accessible:
          type: boolean
          example: false
        price:
          type: number
          minimum: 0
          example: 0
        isFree:
          type: boolean
          description: "Virtual — true si price === 0"
          example: true
        languages:
          type: array
          items:
            type: string
          example: ["es", "en"]
        mapLeft:
          type: number
          description: "% left calculado para posicionar en el SVG"
          example: 72.34
        mapTop:
          type: number
          description: "% top calculado para posicionar en el SVG"
          example: 28.91
        active:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ActivityInput:
      type: object
      required: [name, category, location, zone]
      properties:
        name:
          type: string
          maxLength: 100
          example: "Ruta del Oso"
        description:
          type: string
          maxLength: 2000
        category:
          type: string
          description: "ObjectId de la categoría"
          example: "664a1b2c3d4e5f6a7b8c9d0e"
        location:
          type: object
          required: [coordinates]
          properties:
            coordinates:
              type: array
              items:
                type: number
              minItems: 2
              maxItems: 2
              description: "[longitud, latitud]"
              example: [-4.9902, 43.2547]
        zone:
          type: string
          enum: [oriente, centro, occidente]
        municipality:
          type: string
        images:
          type: array
          items:
            type: string
        accessible:
          type: boolean
          default: false
        price:
          type: number
          minimum: 0
          default: 0
        languages:
          type: array
          items:
            type: string
          default: ["es"]

    # ── Usuario ─────────────────────────────────────────────
    User:
      type: object
      properties:
        _id:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, visitor]
        active:
          type: boolean
        lastLogin:
          type: string
          format: date-time

    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: "JWT válido 15 minutos"
        refreshToken:
          type: string
          description: "JWT válido 7 días"
        user:
          $ref: '#/components/schemas/User'

    # ── Errores ─────────────────────────────────────────────
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Mensaje de error"

    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: "Datos de entrada inválidos"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

# ─── Endpoints ──────────────────────────────────────────────
paths:

  # ── Health ─────────────────────────────────────────────────
  /health:
    get:
      summary: Estado del servidor
      operationId: getHealth
      responses:
        '200':
          description: Servidor operativo
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
                  timestamp:
                    type: string
                    format: date-time
                  db:
                    type: string
                    enum: [connected, disconnected]

  # ── Actividades ────────────────────────────────────────────
  /api/activities:
    get:
      tags: [actividades]
      summary: Listar actividades
      operationId: getActivities
      parameters:
        - name: zone
          in: query
          schema:
            type: string
            enum: [oriente, centro, occidente]
          description: Filtrar por zona geográfica
        - name: category
          in: query
          schema:
            type: string
          description: ObjectId de la categoría
        - name: accessible
          in: query
          schema:
            type: boolean
          description: Solo actividades accesibles
        - name: free
          in: query
          schema:
            type: boolean
          description: Solo actividades gratuitas
        - name: search
          in: query
          schema:
            type: string
            maxLength: 100
          description: Búsqueda fulltext en nombre y descripción
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: Lista paginada de actividades
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Activity'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: Parámetros de filtro inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

    post:
      tags: [actividades]
      summary: Crear actividad
      operationId: createActivity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityInput'
            example:
              name: "Acuario de Gijón"
              description: "El mayor acuario del norte de España."
              category: "664a1b2c3d4e5f6a7b8c9d0e"
              location:
                coordinates: [-5.6618, 43.5454]
              zone: "centro"
              municipality: "Gijón"
              price: 12.5
              accessible: true
              languages: ["es", "en"]
      responses:
        '201':
          description: Actividad creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Token no proporcionado o inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Se requiere rol admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/activities/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: ObjectId de la actividad

    get:
      tags: [actividades]
      summary: Obtener actividad por ID
      operationId: getActivityById
      responses:
        '200':
          description: Actividad encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '400':
          description: ID con formato inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Actividad no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags: [actividades]
      summary: Actualizar actividad
      operationId: updateActivity
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityInput'
      responses:
        '200':
          description: Actividad actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [actividades]
      summary: Desactivar actividad (soft delete)
      operationId: deleteActivity
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Actividad desactivada
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Actividad desactivada correctamente"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Categorías ─────────────────────────────────────────────
  /api/categories:
    get:
      tags: [categorías]
      summary: Listar categorías activas
      operationId: getCategories
      responses:
        '200':
          description: Lista de categorías ordenadas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'

    post:
      tags: [categorías]
      summary: Crear categoría
      operationId: createCategory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        '201':
          description: Categoría creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: El slug ya existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Auth ───────────────────────────────────────────────────
  /api/auth/register:
    post:
      tags: [auth]
      summary: Registrar nuevo usuario
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                  description: "Mínimo 8 caracteres, 1 mayúscula y 1 número"
      responses:
        '201':
          description: Usuario registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: El email ya está registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/auth/login:
    post:
      tags: [auth]
      summary: Iniciar sesión
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Sesión iniciada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/auth/refresh:
    post:
      tags: [auth]
      summary: Renovar access token
      operationId: refreshToken
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Nuevos tokens generados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/auth/me:
    get:
      tags: [auth]
      summary: Obtener usuario autenticado
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Datos del usuario actual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── Respuestas reutilizables ────────────────────────────────
  responses:
    BadRequest:
      description: Datos de entrada inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    Unauthorized:
      description: Token no proporcionado o inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Sin permisos suficientes
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'